pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image-editor"
version = "0.1.0"
description = "A versatile image and photo editing GUI/CLI tool"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
dependencies = [
    "Pillow>=10.0.0",
    "click>=8.0.0",
    "opencv-python-headless>=4.8.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov",
]

[project.scripts]
image-editor = "image_editor.cli:main"
image-editor-gui = "image_editor.gui:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["image_editor*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
